feat/refactor: refactored controller to work with new promise based connection and implemented authorization

+++ backend/src/controllers/wishlist-controller.ts
@@ -1,35 +1,60 @@
 import type { Request, Response } from "express"
 import { connection } from "../utils/db-connection"
-import { handleExists, handleQueryOutput } from "../utils/query-handling"
+import { handleUser } from "../utils/auth";
+import { handleExists } from "../utils/query-handling"
 
 export const addToWishlist = async (req: Request, res: Response) => {
+    const user = await handleUser(req, res, ["user", "admin", "master"]);
+    if (!user) {
+        return;
+    }
+    
     await connection.execute(
         `INSERT INTO wishlist (userId, gameId) VALUES (?, ?)`,
-        [req.params["userId"], req.params["gameId"]],
-        handleQueryOutput(201, res)
+        [user.userId, req.params["gameId"]]
     )
+    res.status(201).send("Game added to wishlist");
 };
 
 export const listWishlistOfUser = async (req: Request, res: Response) => {
-    await connection.execute(
+    const user = await handleUser(req, res, ["user", "admin", "master"]);
+    if (!user) {
+        return;
+    }
+
+    const [wishlist] = await connection.execute(
         `SELECT gameId FROM wishlist WHERE userId = ?`,
-        [req.params["userId"]],
-        handleQueryOutput(200, res)
+        [user.userId]
     )
+    res.status(200).json(wishlist);
 };
 
 export const deleteFromWishlist = async (req: Request, res: Response) => {
+    const user = await handleUser(req, res, ["user", "admin", "master"]);
+    if (!user) {
+        return;
+    }
+
     await connection.execute(
         `DELETE FROM wishlist WHERE userId = ? AND gameId = ?`,
-        [req.params["userId"], req.params["gameId"]],
-        handleQueryOutput(200, res)
+        [user.userId, req.params["gameId"]]
     )
+    res.status(200).send("Game removed from wishlist");
 };
 
 export const isGameInWishlist = async (req: Request, res: Response) => {
-    await connection.execute(
-        `SELECT * FROM wishlist WHERE userId = ? AND gameId = ?`,
-        [req.params["userId"], req.params["gameId"]],
-        handleExists(res)
-    )
+    const user = await handleUser(req, res, ["user", "admin", "master"]);
+    if (!user) {
+        return;
+    }
+
+    handleExists(res, () => checkWishlist(user.userId, Number(req.params["gameId"])));
 };
+
+const checkWishlist = async (userId: number, gameId: number): Promise<boolean> => {
+    const [rows] = await connection.execute(
+        `SELECT * FROM wishlist WHERE userId = ? AND gameId = ?`,
+        [userId, gameId]
+    );
+    return (Array.isArray(rows) && rows.length > 0);
+}
